@@ -23,8 +23,8 @@
     "url": "https://github.com/brunoc107/ngx-viacep.git/issues"
   },
   "peerDependencies": {
-    "@angular/common": ">=14.3.0",
-    "@angular/core": ">=14.3.0"
+    "@angular/common": "^18.0.1",
+    "@angular/core": "^18.0.1"
   },
   "dependencies": {
     "tslib": "^2.3.0"
